Backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
postgres==4.0
djangorestframework-jwt==1.11.0
pyjwt==v1.7.1
django-environ==0.11.2
Pillow==8.4.0